# Core/StickyBoard.Core/Docs/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "StickyBoard API",
    "description": "REST API for StickyBoard cross-platform workspace (Academic Project)",
    "contact": {
      "name": "Alexandre Emond",
      "url": "https://aedev.pro"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/Invites": {
      "post": {
        "tags": [
          "Invites"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteCreateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteCreateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/InviteCreateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InviteCreateResponseDtoApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/InviteCreateResponseDtoApiResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "example": {
                  "code": "VALIDATION_ERROR",
                  "message": "Invalid input",
                  "details": "Additional debug info in development mode only"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "example": {
                  "code": "AUTH_INVALID",
                  "message": "Authentication required or invalid token",
                  "details": "Additional debug info in development mode only"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "example": {
                  "code": "FORBIDDEN",
                  "message": "User does not have access to this resource",
                  "details": "Additional debug info in development mode only"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "example": {
                  "code": "NOT_FOUND",
                  "message": "Requested resource does not exist",
                  "details": "Additional debug info in development mode only"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "example": {
                  "code": "SERVER_ERROR",
                  "message": "Unexpected server error",
                  "details": "Additional debug info in development mode only"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Invites/sent": {
      "get": {
        "tags": [
          "Invites"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InviteListItemDtoIEnumerableApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/InviteListItemDtoIEnumerableApiResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "example": {
                  "code": "VALIDATION_ERROR",
                  "message": "Invalid input",
                  "details": "Additional debug info in development mode only"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "example": {
                  "code": "AUTH_INVALID",
                  "message": "Authentication required or invalid token",
                  "details": "Additional debug info in development mode only"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "example": {
                  "code": "FORBIDDEN",
                  "message": "User does not have access to this resource",
                  "details": "Additional debug info in development mode only"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "example": {
                  "code": "NOT_FOUND",
                  "message": "Requested resource does not exist",
                  "details": "Additional debug info in development mode only"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "example": {
                  "code": "SERVER_ERROR",
                  "message": "Unexpected server error",
                  "details": "Additional debug info in development mode only"
                }
              }
            }
          }
        }
      }
    },
    "/api/Invites/{inviteId}": {
      "delete": {
        "tags": [
          "Invites"
        ],
        "parameters": [
          {
            "name": "inviteId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "example": {
                  "code": "VALIDATION_ERROR",
                  "message": "Invalid input",
                  "details": "Additional debug info in development mode only"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "example": {
                  "code": "AUTH_INVALID",
                  "message": "Authentication required or invalid token",
                  "details": "Additional debug info in development mode only"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "example": {
                  "code": "FORBIDDEN",
                  "message": "User does not have access to this resource",
                  "details": "Additional debug info in development mode only"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "example": {
                  "code": "NOT_FOUND",
                  "message": "Requested resource does not exist",
                  "details": "Additional debug info in development mode only"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "example": {
                  "code": "SERVER_ERROR",
                  "message": "Unexpected server error",
                  "details": "Additional debug info in development mode only"
                }
              }
            }
          }
        }
      }
    },
    "/api/Invites/public/{token}": {
      "get": {
        "tags": [
          "Invites"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitePublicDtoApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitePublicDtoApiResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "example": {
                  "code": "VALIDATION_ERROR",
                  "message": "Invalid input",
                  "details": "Additional debug info in development mode only"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "example": {
                  "code": "AUTH_INVALID",
                  "message": "Authentication required or invalid token",
                  "details": "Additional debug info in development mode only"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "example": {
                  "code": "FORBIDDEN",
                  "message": "User does not have access to this resource",
                  "details": "Additional debug info in development mode only"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "example": {
                  "code": "NOT_FOUND",
                  "message": "Requested resource does not exist",
                  "details": "Additional debug info in development mode only"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "example": {
                  "code": "SERVER_ERROR",
                  "message": "Unexpected server error",
                  "details": "Additional debug info in development mode only"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BoardRole": {
        "enum": [
          0,
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "ErrorCode": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5
        ],
        "type": "integer",
        "format": "int32"
      },
      "ErrorDto": {
        "type": "object",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "details": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "InviteCreateDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          },
          "boardId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "orgId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "boardRole": {
            "$ref": "#/components/schemas/BoardRole"
          },
          "orgRole": {
            "$ref": "#/components/schemas/OrgRole"
          },
          "expiresInDays": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "InviteCreateResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "token": {
            "type": "string",
            "nullable": true
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "InviteCreateResponseDtoApiResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "$ref": "#/components/schemas/InviteCreateResponseDto"
          }
        },
        "additionalProperties": false
      },
      "InviteListItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "boardId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "organizationId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "boardRole": {
            "$ref": "#/components/schemas/BoardRole"
          },
          "orgRole": {
            "$ref": "#/components/schemas/OrgRole"
          },
          "accepted": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "senderDisplayName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "InviteListItemDtoIEnumerableApiResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InviteListItemDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "InvitePublicDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          },
          "boardId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "organizationId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "boardRole": {
            "$ref": "#/components/schemas/BoardRole"
          },
          "orgRole": {
            "$ref": "#/components/schemas/OrgRole"
          },
          "accepted": {
            "type": "boolean"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "senderDisplayName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "InvitePublicDtoApiResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "$ref": "#/components/schemas/InvitePublicDto"
          }
        },
        "additionalProperties": false
      },
...
      "ObjectApiResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "OrgRole": {
        "enum": [
          0,
          1,
          2,
          3,
          4
        ],
        "type": "integer",
        "format": "int32"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT Authorization (use: Bearer <token>)",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "ApiKey": {
        "type": "apiKey",
        "description": "Worker API Key (use: ApiKey <your_key>)",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [],
      "ApiKey": []
    }
  ]
}